% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactrates.R
\name{con1}
\alias{con1}
\alias{con2}
\alias{con3}
\title{Contact Rates}
\usage{
con1(i, p, r, nc, o, uh, uo)

con2(i, p, r, nc, o, uh, uo, e = 1)

con3(i, p, r, nc, o)
}
\arguments{
\item{i}{The total number of completed interviews.}

\item{p}{The total number of partial interviews.}

\item{r}{The total number of refusals.}

\item{nc}{The total number of non-contacts.}

\item{o}{The total number of \dQuote{other} units.}

\item{uh}{The total number of \emph{unknown if occupied} (housing) units.}

\item{uo}{The total number of other unknowns.}

\item{e}{Optionally, the (estimated) proportion of units of unknown eligibility (\code{uh} and\code{uo}) that are eligible. Otherwise assumed to be 1 (all eligible).}
}
\value{
A numeric value indicating the contact rate.
}
\description{
AAPOR contact rate formulae
}
\seealso{
\code{\link{rr1}}, \code{\link{coop1}}, \code{\link{ref1}}
}

