% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooperationrates.R
\name{coop1}
\alias{coop1}
\alias{coop2}
\alias{coop3}
\alias{coop4}
\title{Cooperation Rates}
\usage{
coop1(i = 0, p = 0, r = 0, o = 0)

coop2(i = 0, p, r = 0, o = 0)

coop3(i = 0, p = 0, r = 0)

coop4(i = 0, p = 0, r = 0)
}
\arguments{
\item{i}{The total number of completed interviews.}

\item{p}{The total number of partial interviews.}

\item{r}{The total number of refusals.}

\item{o}{The total number of \dQuote{other} units.}
}
\value{
A numeric value indicating the cooperation rate.
}
\description{
AAPOR cooperation rate formulae
}
\references{
AAPOR. 2016. \href{http://www.aapor.org/AAPOR_Main/media/publications/Standard-Definitions20169theditionfinal.pdf}{\dQuote{Standard Definitions Final Dispositions of Case Codes and Outcome Rates for Surveys.}}
}
\seealso{
\code{\link{rates}}, \code{\link{rr1}}, \code{\link{ref1}}, \code{\link{con1}}
}

