% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{rates}
\alias{rates}
\alias{rates.character}
\alias{rates.factor}
\alias{rates.list}
\alias{rates.table}
\title{AAPOR Survey Response Rates}
\usage{
rates(x, ...)

\method{rates}{factor}(x, ...)

\method{rates}{character}(x, w = NULL, e = 1, ...)

\method{rates}{table}(x, ...)

\method{rates}{list}(x, e = 1, ...)
}
\arguments{
\item{x}{A character or factor vector containing unit-specific disposition codes, or a named table or list summarizing the disposition codes.}

\item{w}{Optionally, if \code{x} is a vector of disposition codes, the weights to be applied to each case in calculating final, weighted disposition counts.}

\item{e}{Optionally, the (estimated) proportion of units of unknown eligibility (\code{uh} and\code{uo}) that are eligible. Otherwise assumed to be 1 (all eligible).}

\item{\dots}{Additional arguments passed to methods.}
}
\value{
A list of class \dQuote{aapor_rates} containing the calculated response rates. Printing of the results is handled by a method that accepts a \code{fmt} argument in the form required by \code{\link[base]{sprintf}}.
}
\description{
Calculate AAPOR response, cooperation, refusal, and contact rates.
}
\details{
This function calculates all response, cooperation, refusal, and contact rates based upon the disposition code information supplied in \code{x}.
The allowed disposition codes are \dQuote{i}, \dQuote{p}, \dQuote{r}, \dQuote{nc}, \dQuote{o}, \dQuote{uh}, and \dQuote{uo}. Both upper and lower case are allowed.
\code{e} is assumed to be 1 unless specified otherwise. Smith (2009) provides a review of estimation of \code{e}.
}
\examples{
# `x` is a character vector of codes
rates(c(rep("I", 100), rep("NC", 20)))
rates(c(rep("I", 100), rep("NC", 20)), w = runif(120))

# `x` is a list/table of code counts
rates(list(i = 100, nc = 20))

# specifying `e`
rates(list(i = 100, nc = 20, uo = 100, e = .25))

}
\references{
AAPOR. 2016. \href{http://www.aapor.org/AAPOR_Main/media/publications/Standard-Definitions20169theditionfinal.pdf}{\dQuote{Standard Definitions Final Dispositions of Case Codes and Outcome Rates for Surveys.}}
 Smith, Tom W. 2009. \href{http://www.aapor.org/AAPOR_Main/media/MainSiteFiles/ERATE09.pdf}{\dQuote{A Revised Review of Methods to Estimate the Status of Cases with Unknown Eligibility.}} NORC/University of Chicago.
}
\seealso{
See \code{\link{rr1}}, \code{\link{ref1}}, \code{\link{coop1}}, \code{\link{con1}} for calculation of each rate.
}

