% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refusalrates.R
\name{ref1}
\alias{ref1}
\alias{ref2}
\alias{ref3}
\title{Refusal Rates}
\usage{
ref1(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0)

ref2(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0, e = 1)

ref3(i = 0, p = 0, r = 0, nc = 0, o = 0)
}
\arguments{
\item{i}{The total number of completed interviews.}

\item{p}{The total number of partial interviews.}

\item{r}{The total number of refusals.}

\item{nc}{The total number of non-contacts.}

\item{o}{The total number of \dQuote{other} units.}

\item{uh}{The total number of \emph{unknown if occupied} (housing) units.}

\item{uo}{The total number of other unknowns.}

\item{e}{Optionally, the (estimated) proportion of units of unknown eligibility (\code{uh} and\code{uo}) that are eligible. Otherwise assumed to be 1 (all eligible).}
}
\value{
A numeric value indicating the refusal rate.
}
\description{
AAPOR refusal rate formulae
}
\details{
Refusal rates measure the number of refused interviews out of the total units sampled. The denominator for each rate varies. Here are the definitions of each response rate, according to AAPOR's \dQuote{Standard Definitions}:
\itemize{
 \item Refusal Rate 1 (REF1) is the number of refusals divided by the interviews (complete and partial) plus the non-respondents (refusals, non-contacts, and others) plus the cases of unknown eligibility.
 \item Refusal Rate 2 (REF2) includes estimated eligible cases among the unknown cases similar to Response Rate 3 (RR3) and Response Rate 4 (RR4).
 \item Refusal Rate 3 is analogous to Response Rate 5 (RR5) and Response Rate 6 (RR6). As in those cases the elimination of the unknowns from the equation must be fully justified by the actual situation. Non-contact and other rates can be calculated in a manner similar to refusal rates. Refusal, non-contact, and other rates will sum to equal the non-response rate.
}
One possible calculation of \code{e} suggested by the Standard Definition's calculator is the number of eligible \eqn{e = \frac{i + p + r + nc + o}{i + p + r + nc + o + uo}}. The default used here, however, is the more conservative value of 1.
}
\references{
AAPOR. 2016. \href{http://www.aapor.org/AAPOR_Main/media/publications/Standard-Definitions20169theditionfinal.pdf}{\dQuote{Standard Definitions Final Dispositions of Case Codes and Outcome Rates for Surveys.}}
}
\seealso{
\code{\link{rates}}, \code{\link{rr1}}, \code{\link{coop1}}, \code{\link{con1}}
}

