% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responserates-package.R
\docType{package}
\name{responserates-package}
\alias{responserates}
\alias{responserates-package}
\title{responserates}
\description{
AAPOR Survey Response Rates
}
\details{
Functions to calculate survey response rates according to the American Association for Public Opinion Research's (AAPOR) Standard Definitions.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\url{http://www.aapor.org/Standards-Ethics/Standard-Definitions-(1).aspx}
}
\keyword{package}

