% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactrates.R
\name{con1}
\alias{con1}
\alias{con2}
\alias{con3}
\title{Contact Rates}
\usage{
con1(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0)

con2(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0, e = 1)

con3(i = 0, p = 0, r = 0, nc = 0, o = 0)
}
\arguments{
\item{i}{The total number of completed interviews.}

\item{p}{The total number of partial interviews.}

\item{r}{The total number of refusals.}

\item{nc}{The total number of non-contacts.}

\item{o}{The total number of \dQuote{other} units.}

\item{uh}{The total number of \emph{unknown if occupied} (housing) units.}

\item{uo}{The total number of other unknowns.}

\item{e}{Optionally, the (estimated) proportion of units of unknown eligibility (\code{uh} and\code{uo}) that are eligible. Otherwise assumed to be 1 (all eligible).}
}
\value{
A numeric value indicating the contact rate.
}
\description{
AAPOR contact rate formulae
}
\references{
AAPOR. 2016. \href{http://www.aapor.org/AAPOR_Main/media/publications/Standard-Definitions20169theditionfinal.pdf}{\dQuote{Standard Definitions Final Dispositions of Case Codes and Outcome Rates for Surveys.}}
}
\seealso{
\code{\link{rates}}, \code{\link{rr1}}, \code{\link{coop1}}, \code{\link{ref1}}
}

