% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooperationrates.R
\name{coop1}
\alias{coop1}
\alias{coop2}
\alias{coop3}
\alias{coop4}
\title{Cooperation Rates}
\usage{
coop1(i, p, r, o)

coop2(i, p, r, o)

coop3(i, p, r)

coop4(i, p, r)
}
\arguments{
\item{i}{The total number of completed interviews.}

\item{p}{The total number of partial interviews.}

\item{r}{The total number of refusals.}

\item{o}{The total number of \dQuote{other} units.}
}
\value{
A numeric value indicating the cooperation rate.
}
\description{
AAPOR cooperation rate formulae
}
\seealso{
\code{\link{rr1}}, \code{\link{ref1}}, \code{\link{con1}}
}

