% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responserates.R
\name{rr1}
\alias{rr1}
\alias{rr2}
\alias{rr3}
\alias{rr4}
\alias{rr5}
\alias{rr6}
\title{Response Rates}
\usage{
rr1(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0)

rr2(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0)

rr3(i = 0, p = 0, r = 0, nc, o = 0, uh = 0, uo = 0, e = 1)

rr4(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0, e = 1)

rr5(i = 0, p = 0, r = 0, nc = 0, o = 0)

rr6(i = 0, p = 0, r = 0, nc = 0, o = 0)
}
\arguments{
\item{i}{The total number of completed interviews.}

\item{p}{The total number of partial interviews.}

\item{r}{The total number of refusals.}

\item{nc}{The total number of non-contacts.}

\item{o}{The total number of \dQuote{other} units.}

\item{uh}{The total number of \emph{unknown if occupied} (housing) units.}

\item{uo}{The total number of other unknowns.}

\item{e}{Optionally, the (estimated) proportion of units of unknown eligibility (\code{uh} and\code{uo}) that are eligible. Otherwise assumed to be 1 (all eligible).}
}
\value{
A numeric value indicating the response rate.
}
\description{
AAPOR response rate formulae
}
\references{
AAPOR. 2016. \href{http://www.aapor.org/AAPOR_Main/media/publications/Standard-Definitions20169theditionfinal.pdf}{\dQuote{Standard Definitions Final Dispositions of Case Codes and Outcome Rates for Surveys.}}
}
\seealso{
\code{\link{rates}}, \code{\link{coop1}}, \code{\link{ref1}}, \code{\link{con1}}
}

