% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responserates.R
\name{rr1}
\alias{rr1}
\alias{rr2}
\alias{rr3}
\alias{rr4}
\alias{rr5}
\alias{rr6}
\title{Response Rates}
\usage{
rr1(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0)

rr2(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0)

rr3(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0, e = 1)

rr4(i = 0, p = 0, r = 0, nc = 0, o = 0, uh = 0, uo = 0, e = 1)

rr5(i = 0, p = 0, r = 0, nc = 0, o = 0)

rr6(i = 0, p = 0, r = 0, nc = 0, o = 0)
}
\arguments{
\item{i}{The total number of completed interviews.}

\item{p}{The total number of partial interviews.}

\item{r}{The total number of refusals.}

\item{nc}{The total number of non-contacts.}

\item{o}{The total number of \dQuote{other} units.}

\item{uh}{The total number of \emph{unknown if occupied} (housing) units.}

\item{uo}{The total number of other unknowns.}

\item{e}{Optionally, the (estimated) proportion of units of unknown eligibility (\code{uh} and\code{uo}) that are eligible. Otherwise assumed to be 1 (all eligible).}
}
\value{
A numeric value indicating the response rate.
}
\description{
AAPOR response rate formulae
}
\details{
Response rates measure the number of complete (or complete and partial) interviews out of the total units sampled. The denominator for each rate varies. Here are the definitions of each response rate, according to AAPOR's \dQuote{Standard Definitions}:
\itemize{
 \item Response Rate 1 (RR1), or the minimum response rate, is the number of complete interviews divided by the number of interviews (complete plus partial) plus the number of non-interviews (refusal and break-off plus non-contacts plus others) plus all cases of unknown eligibility (unknown if housing unit, plus unknown, other).
 \item Response Rate 2 (RR2) counts partial interviews as respondents.
 \item Response Rate 3 (RR3) estimates what proportion of cases of unknown eligibility is actually eligible. In estimating \code{e}, one must be guided by the best available scientific information on what share eligible cases make up among the unknown cases and one must not select a proportion in order to boost the response rate.
 \item Response Rate 4 (RR4) allocates cases of unknown eligibility as in RR3, but also includes partial interviews as respondents as in RR2.
 \item Response Rate 5 (RR5) is either a special case of RR3 in that it assumes that \samp{e=0} (i.e. that there are no eligible cases among the cases of unknown eligibility) or the rare case in which there are no cases of unknown eligibility
 \item Response Rate 6 (RR6) makes that same assumption and also includes partial interviews as respondents. RR6 represents the maximum response rate.
}
One possible calculation of \code{e} suggested by the Standard Definition's calculator is the number of eligible \eqn{e = \frac{i + p + r + nc + o}{i + p + r + nc + o + uo}}. The default used here, however, is the more conservative value of 1.
}
\references{
AAPOR. 2016. \href{http://www.aapor.org/AAPOR_Main/media/publications/Standard-Definitions20169theditionfinal.pdf}{\dQuote{Standard Definitions Final Dispositions of Case Codes and Outcome Rates for Surveys.}}
}
\seealso{
\code{\link{rates}}, \code{\link{coop1}}, \code{\link{ref1}}, \code{\link{con1}}
}

